Refactor handles/src/homoglyphs/canonical.rs -> handles/src/utils/converter.rs
