--- pallets/handles/src/homoglyphs/mod.rs
@@ -1,2 +1,0 @@
-pub mod canonical;
-mod confusables;
