+++ pallets/handles/src/utils/confusable_characters.txt
