+++ pallets/handles/src/utils/confusables.rs
