+++ pallets/handles/src/utils/confusables_generator.rs
